+++ Packages/com.lam0uette.uniject/Runtime/Lifecycle/IFixedTickable.cs
@@ -1,0 +1,7 @@
+namespace LaM0uette.UniJect
+{
+    public interface IFixedTickable
+    {
+        void FixedTick(float fixedDeltaTime);
+    }
+}