+++ Packages/com.lam0uette.uniject/Runtime/Lifecycle/ILateTickable.cs
@@ -1,0 +1,7 @@
+namespace LaM0uette.UniJect
+{
+    public interface ILateTickable
+    {
+        void LateTick(float deltaTime);
+    }
+}