+++ Packages/com.lam0uette.uniject/Runtime/Lifecycle/ITickRegistry.cs
@@ -1,0 +1,9 @@
+namespace LaM0uette.UniJect
+{
+    public interface ITickRegistry
+    {
+        void Add(DIContainer container);
+
+        void Remove(DIContainer container);
+    }
+}