+++ Packages/com.lam0uette.uniject/Runtime/Unity/Lifecycle/TickPhase.cs
@@ -1,0 +1,6 @@
+namespace LaM0uette.UniJect
+{
+    internal struct TickPhase
+    {
+    }
+}